Update README.md
A subscriber to a channel on the video-sharing YouTube is a user who has chosen to receive the channel's content by clicking on that channel's "Subscribe" button, and each user's subscription feed consists of videos published by channels to which the user is subscribed. The ability to subscribe to users was introduced in October 2005. YouTube began publishing a list of its most- subscribed channels in April 2006. An early archive of the list dates to May 2006.

The following table lists the 50 most-subscribed YouTube channels, as well as the primary language and content category of each channel. The channels are ordered by number of subscribers; those whose displayed subscriber counts are identical are listed so that the channel whose current growth rate indicates that its displayed subscriber count will exceed that of the other channel is listed first. Automatically generated channels that lack their own videos (such as Music and News) and channels that have been made effectively obsolete as a result of the transferal of their content (such as Justin Bieber VEVO and Taylor Swift VEVO) are excluded. As of February 2023, 21 of the 50 channels listed primarily produce content in English while 16 primarily produce content in Hindi. All 50 of the channels have surpassed 40 million subscribers, 39 of them have surpassed 50 million subscribers, 23 of them have surpassed 60 million subscribers, 16 of them have surpassed 70 million subscribers, 12 of them have surpassed 80 million subscribers, 10 of them have surpassed 90 million and 7 of them have surpassed 100 million subscribers. Only 1 channel (T-Series) has surpassed 200 million subscribers.